Organize classes into java packages
