--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-public abstract class Command {
-    public abstract void execute(Ui ui, TaskList taskList, Storage storage) throws DukeException;
-    public boolean isBye() {
-        return false;
-    }
-}
