+++ src/main/java/Duke.java
@@ -1,3 +1,11 @@
+import duke.command.Command;
+import duke.command.SetupCommand;
+import duke.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class Duke {
     public static void main(String[] args) {
         Storage storage = new Storage();
