--- src/main/java/DukeException.java
@@ -1,5 +1,0 @@
-public class DukeException extends Exception {
-    public DukeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
