--- src/main/java/ToDo.java
@@ -1,15 +1,0 @@
-public class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toFileFormat() {
-        return "T," + super.toFileFormat() + ",," + getStatusIcon();    
-    }
-    
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
