+++ src/main/java/duke/command/AddDeadlineTaskCommand.java
@@ -1,0 +1,27 @@
+package duke.command;
+
+import java.time.LocalDate;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class AddDeadlineTaskCommand extends Command {
+    private final String description;
+    private final LocalDate date;
+
+    public AddDeadlineTaskCommand(String description, LocalDate date) {
+        this.description = description;
+        this.date = date;
+    }
+
+    @Override
+    public void execute(Ui ui, TaskList taskList, Storage storage) {
+        taskList.addDeadlineTask(description, date);
+
+        int taskIndex = taskList.getNumberOfTasks() - 1;
+        ui.showMessage("Got it. I've added this task: ");
+        ui.showMessage(taskList.getDescriptionOfTaskAtIndex(taskIndex));
+        ui.showMessage("Now you have " + taskList.getNumberOfTasks() + " tasks in the list.");
+    }
+}
