+++ src/main/java/duke/command/AddEventTaskCommand.java
@@ -1,0 +1,25 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class AddEventTaskCommand extends Command {
+    private final String description;
+    private final String at;
+    
+    public AddEventTaskCommand(String description, String at) {
+        this.description = description;
+        this.at = at;
+    }
+
+    @Override
+    public void execute(Ui ui, TaskList taskList, Storage storage) {
+        taskList.addEventTask(description, at);
+
+        int taskIndex = taskList.getNumberOfTasks() - 1;
+        ui.showMessage("Got it. I've added this task: ");
+        ui.showMessage(taskList.getDescriptionOfTaskAtIndex(taskIndex));
+        ui.showMessage("Now you have " + taskList.getNumberOfTasks() + " tasks in the list.");
+    }
+}
