+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class DukeException extends Exception {
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
