+++ src/main/java/duke/task/Event.java
@@ -1,0 +1,22 @@
+package duke.task;
+
+import duke.task.Task;
+
+public class Event extends Task {
+    private String at;
+
+    public Event(String description, String at) {
+        super(description);
+        this.at = at;
+    }
+
+    @Override
+    public String toFileFormat() {
+        return "E," + super.toFileFormat() + "," + at + "," + getStatusIcon();
+    }
+    
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (at: " + at + ")";
+    }
+}
