+++ src/main/java/duke/task/Task.java
@@ -1,0 +1,32 @@
+package duke.task;
+
+public class Task {
+    private String description;
+    private boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    protected String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark task that is done with X
+    }
+
+    public void mark() {
+        isDone = true;
+    }
+
+    public void unmark() {
+        isDone = false;
+    }
+
+    public String toFileFormat() {
+        return description;
+    }
+    
+    @Override
+    public String toString() {
+        return "[" + getStatusIcon() + "] " + description;
+    }
+}
