+++ src/main/java/duke/task/ToDo.java
@@ -1,0 +1,19 @@
+package duke.task;
+
+import duke.task.Task;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toFileFormat() {
+        return "T," + super.toFileFormat() + ",," + getStatusIcon();    
+    }
+    
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
